+++ src/components/Auth/LoginForm.tsx
@@ -2,12 +2,13 @@ import { useReactiveVar } from "@apollo/client";
 import { Card, CardContent, FormGroup } from "@mui/material";
 import { Form, Formik } from "formik";
 import { useTranslation } from "react-i18next";
+import { useLoginMutation } from "../../apollo/auth/mutations/gen/Login.mutation";
 import {
   isLoggedInVar,
   isNavDrawerOpenVar,
   toastVar,
 } from "../../apollo/cache";
-import { LoginInput, useLoginMutation } from "../../apollo/gen";
+import { LoginInput } from "../../apollo/gen";
 import Flex from "../../components/Shared/Flex";
 import LevelOneHeading from "../../components/Shared/LevelOneHeading";
 import PrimaryActionButton from "../../components/Shared/PrimaryActionButton";
