+++ src/pages/auth/signup.tsx
@@ -5,7 +5,8 @@ import { useRouter } from "next/router";
 import { useEffect } from "react";
 import { useTranslation } from "react-i18next";
 import { inviteTokenVar, isLoggedInVar } from "../../apollo/cache";
-import { useIsFirstUserQuery, useServerInviteQuery } from "../../apollo/gen";
+import { useServerInviteQuery } from "../../apollo/invites/queries/gen/ServerInvite.query";
+import { useIsFirstUserQuery } from "../../apollo/users/queries/gen/IsFirstUser.query";
 import SignUpForm from "../../components/Auth/SignUpForm";
 import ProgressBar from "../../components/Shared/ProgressBar";
 import { NavigationPaths } from "../../constants/common.constants";
