+++ src/pages/events/index.tsx
@@ -12,7 +12,8 @@ import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { isLoggedInVar } from "../../apollo/cache";
-import { EventsInput, useEventsLazyQuery } from "../../apollo/gen";
+import { useEventsLazyQuery } from "../../apollo/events/queries/gen/Events.query";
+import { EventsInput } from "../../apollo/gen";
 import EventCompact from "../../components/Events/EventCompact";
 import LevelOneHeading from "../../components/Shared/LevelOneHeading";
 import ProgressBar from "../../components/Shared/ProgressBar";
